REST API finished, socketio, User model, connected to Post model

+++ 19-PracticalREST/server/controllers/feed.js
@@ -2,34 +2,62 @@ const { validationResult } = require('express-validator');
 const Post = require('../models/post');
 const fs = require('fs');
 const path = require('path');
+const User = require('../models/user');
 
-exports.getPosts = (req, res, next) => {
-    // implement pagination
+const io = require('../socket');
+const user = require('../models/user');
+
+// exports.getPosts = (req, res, next) => {
+//     // implement pagination
+//     const currentPage = req.query.page || 1;
+//     // posts per page
+//     const perPage = 2;
+//     let totalItems;
+//     Post.find()
+//         .countDocuments()
+//         .then((count) => {
+//             totalItems = count;
+//             return Post.find()
+//                 .skip((currentPage - 1) * perPage)
+//                 .limit(perPage)
+//                 .then((posts) => {
+//                     res.status(200).json({
+//                         message: 'Fetched posts successfully',
+//                         posts: posts,
+//                         totalItems: totalItems,
+//                     });
+//                 });
+//         })
+//         .catch((err) => {
+//             if (!err.statusCode) {
+//                 err.statusCode = 500;
+//             }
+//             next(err);
+//         });
+// };
+
+exports.getPosts = async (req, res, next) => {
     const currentPage = req.query.page || 1;
-    // posts per page
     const perPage = 2;
-    let totalItems;
-    Post.find()
-        .countDocuments()
-        .then((count) => {
-            totalItems = count;
-            return Post.find()
-                .skip((currentPage - 1) * perPage)
-                .limit(perPage)
-                .then((posts) => {
-                    res.status(200).json({
-                        message: 'Fetched posts successfully',
-                        posts: posts,
-                        totalItems: totalItems,
-                    });
-                });
-        })
-        .catch((err) => {
-            if (!err.statusCode) {
-                err.statusCode = 500;
-            }
-            next(err);
+    try {
+        let totalItems = await Post.find().countDocuments();
+        const posts = await Post.find()
+            .populate('creator')
+            .sort({ createdAt: -1 })
+            .skip((currentPage - 1) * perPage)
+            .limit(perPage);
+
+        res.status(200).json({
+            message: 'Fetched posts successfully',
+            posts: posts,
+            totalItems: totalItems,
         });
+    } catch (err) {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    }
 };
 
 exports.createPost = (req, res, next) => {
@@ -56,6 +84,7 @@ exports.createPost = (req, res, next) => {
     //// MULTER WARNING
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
 
     const post = new Post({
         // in the model we asked mongoose to create these for us
@@ -63,16 +92,32 @@ exports.createPost = (req, res, next) => {
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: { name: 'Mike' },
+        creator: req.userId,
         // createdAt: new Date(),
     });
     // .save() returns a promise so you get access to the result in the next then() block
     post.save()
         .then((result) => {
-            console.log(result);
+            return User.findById(req.userId);
+        })
+        .then((user) => {
+            creator = user;
+            user.posts.push(post);
+            return user.save();
+        })
+        .then((result) => {
+            // emit sends to all users, broadcast sends to all users but this.
+            io.getIO().emit('posts', {
+                action: 'create',
+                post: {
+                    ...post._doc,
+                    creator: { _id: req.userId, name: user.name },
+                },
+            });
             res.status(201).json({
                 message: 'Post created successfully!',
-                post: result,
+                post: post,
+                creator: { _id: creator._id, name: creator.name },
             });
         })
         // this is tied to the "throw err" above
@@ -135,6 +180,7 @@ exports.updatePost = (req, res, next) => {
     }
     // now we know the data is valid. Now we update.
     Post.findById(postId)
+        .populate('creator')
         .then((post) => {
             if (!post) {
                 const error = new Error('Could not find post');
@@ -142,6 +188,12 @@ exports.updatePost = (req, res, next) => {
                 // throw to catch it
                 throw err;
             }
+            if (post.creator._id.toString() !== req.userId) {
+                const err = new Error('Not authorized');
+                err.statusCode = 403;
+                throw err;
+            }
+
             // checking for imageURL change and then deleting the old imageUrl
             if (imageUrl !== post.imageUrl) {
                 clearImage(post.imageUrl);
@@ -153,6 +205,7 @@ exports.updatePost = (req, res, next) => {
             return post.save();
         })
         .then((result) => {
+            io.getIO.emit('posts', { action: 'update', post: result });
             // return data to client in a response.
             res.status(200).json({ message: 'Post updated', post: result });
         })
@@ -175,11 +228,24 @@ exports.deletePost = (req, res, next) => {
                 // throw to catch it
                 throw err;
             }
+            if (post.creator.toString() !== req.userId) {
+                const err = new Error('Not authorized');
+                err.statusCode = 403;
+                throw err;
+            }
             // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndRemove(postId);
         })
         .then((result) => {
+            return User.findById(req.userId);
+        })
+        .then((user) => {
+            user.posts.pull(postId);
+            return user.save();
+        })
+        .then((result) => {
+            io.getIO().emit('posts', { action: 'delete', post: postId });
             console.log(result);
             res.status(200).json({ message: 'Deleted Post ' + postId });
         })
